refactor:MainScreen 대신 HomeScreen으로 용어 및 컴포넌트 통일
